backend/requirements.txt
alembic==1.12.1
annotated-types==0.5.0
anyio==3.7.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.3.0
email-validator==2.0.0.post2
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.103.2
greenlet==3.1.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
jinja2==3.1.4
Mako==1.2.4
MarkupSafe==2.1.5
mysqlclient==2.1.1
numpy==1.21.6
openpyxl==3.1.3
orjson==3.9.7
pandas==1.3.5
psycopg2==2.9.9
pydantic==2.5.3
pydantic-core==2.14.6
pydantic-extra-types==2.3.0
pydantic-settings==2.0.3
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-multipart==0.0.8
pytz==2024.2
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.27.0
typing-extensions==4.7.1
ujson==5.7.0
urllib3==2.0.7
uvicorn==0.22.0
uvloop==0.18.0
watchfiles==0.20.0
websockets==11.0.3
zipp==3.15.0
